Version 1.3.4 Update dependencies and optimize image processing
Removed imgscalr-lib and added simple-png-javafx as a dependency to enhance image processing and remove dependency on AWT and Swing. Optimized the image resizing process in 'ProcessFile' by reusing the last resized image when the next image size is the same as the last one. Unnecessary print statements were removed, and project version was updated to 1.3.4.